Revert "feat: Prise en charge Mermaid dans BlockNote"
