+++ frontend/src/components/blog/PostDetail.jsx
@@ -1,13 +1,54 @@
+import "@blocknote/core/fonts/inter.css";
+import "@blocknote/mantine/style.css";
+import { useCreateBlockNote } from "@blocknote/react";
 import { Alert } from "@mantine/core";
-import { useEffect, useState } from "react";
+import DOMPurify from "dompurify";
+import { useEffect, useMemo, useState } from "react";
 import { Helmet } from "react-helmet-async";
 import { Link, useParams } from "react-router-dom";
 import { api } from "../../api/client";
 import { useAuth } from "../../contexts/AuthContext";
 import Avatar from "../ui/Avatar";
-import BlockNoteRenderer from "./BlockNoteRenderer";
 import CommentForm from "./CommentForm";
 import CommentSection from "./CommentSection";
+
+function BlockNoteRenderer({ content }) {
+  const blocks = useMemo(() => {
+    try {
+      return JSON.parse(content);
+    } catch {
+      return null;
+    }
+  }, [content]);
+
+  const validBlocks =
+    Array.isArray(blocks) && blocks.length > 0 ? blocks : undefined;
+
+  const editor = useCreateBlockNote({
+    initialContent: validBlocks,
+  });
+
+  const [html, setHtml] = useState("");
+
+  useEffect(() => {
+    if (editor && validBlocks) {
+      editor.blocksToFullHTML(editor.document).then((rawHtml) => {
+        setHtml(DOMPurify.sanitize(rawHtml));
+      }).catch(() => {});
+    }
+  }, [editor, validBlocks]);
+
+  if (!validBlocks) {
+    return <div className="whitespace-pre-line">{content}</div>;
+  }
+
+  return (
+    <div
+      className="text-gray-700 leading-relaxed"
+      dangerouslySetInnerHTML={{ __html: html }}
+    />
+  );
+}
 
 export default function PostDetail() {
   const { slug } = useParams();
@@ -213,7 +254,7 @@ export default function PostDetail() {
           </p>
         </div>
 
-        <BlockNoteRenderer key={post.content} content={post.content} />
+        <BlockNoteRenderer content={post.content} />
       </article>
 
       <CommentSection
